scripts/requirements.txt
pandas==2.3.3
numpy==2.3.3
gspread==6.2.1
google-auth==2.41.1
google-api-python-client==2.184.0